feat(cli): add default plugin folder for MAC OS
closes #5806
